backend/fileconverter/textures.py
"""
Texture and model format conversion utilities
"""
from pathlib import Path
from typing import Optional, Union, Literal
from PIL import Image

class TextureConverter:
    """Handles conversion between different texture formats"""
    
    @staticmethod
    def dds_to_png(
        input_file: Union[str, Path],
        output_file: Optional[Union[str, Path]] = None
    ) -> Path:
        """
        Convert DDS texture to PNG format
        
        Args:
            input_file: Path to input DDS file
            output_file: Optional output PNG file path
            
        Returns:
            Path to created PNG file
        """
        input_path = Path(input_file)
        if not input_path.exists():
            raise FileNotFoundError(f"Input file not found: {input_path}")
            
        if output_file is None:
            output_file = input_path.with_suffix('.png')
            
        # Implementation using Pillow or similar
        raise NotImplementedError("DDS to PNG conversion not yet implemented")

    @staticmethod
    def png_to_dds(
        input_file: Union[str, Path],
        output_file: Optional[Union[str, Path]] = None,
        format: Literal["DXT1", "DXT3", "DXT5"] = "DXT5"
    ) -> Path:
        """
        Convert PNG to DDS texture format
        
        Args:
            input_file: Path to input PNG file
            output_file: Optional output DDS file path
            format: DDS compression format to use
            
        Returns:
            Path to created DDS file
        """
        input_path = Path(input_file)
        if not input_path.exists():
            raise FileNotFoundError(f"Input file not found: {input_path}")
            
        if output_file is None:
            output_file = input_path.with_suffix('.dds')
            
        # Implementation using Pillow or similar
        raise NotImplementedError("PNG to DDS conversion not yet implemented")

class ModelConverter:
    """Handles conversion between different 3D model formats"""
    
    @staticmethod
    def obj_to_fbx(
        input_file: Union[str, Path],
        output_file: Optional[Union[str, Path]] = None
    ) -> Path:
        """
        Convert OBJ model to FBX format
        
        Args:
            input_file: Path to input OBJ file
            output_file: Optional output FBX file path
            
        Returns:
            Path to created FBX file
        """
        input_path = Path(input_file)
        if not input_path.exists():
            raise FileNotFoundError(f"Input file not found: {input_path}")
            
        if output_file is None:
            output_file = input_path.with_suffix('.fbx')
            
        # Implementation using Blender Python API or similar
        raise NotImplementedError("OBJ to FBX conversion not yet implemented") 