backend/pyproject.toml
[tool.poetry]
name = "fileconverter"
version = "0.1.0"
description = "A comprehensive file conversion utility for game modding and related tasks"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fileconverter"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"  # For image conversions
pydub = "^0.25.1"  # For audio conversions
py7zr = "^0.20.0"  # For 7z compression
rarfile = "^4.0"  # For RAR files
xmltodict = "^0.13.0"  # For XML conversions
pyyaml = "^6.0"  # For YAML conversions

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 